date: drop ntp hack and replace it with different test
/etc/init.d/ntpdate hack does not work now. I tried to implement it
better but problem is that ntpdate is no longer present and replaced by
ntpd. We can set time with ntpd but problem is that it does not report
information about NTP peer, which was intention of this.

Instead of that I added configuration print, forced sync and check after
such sync.