certs: drop
We do not use this at the moment and it is questionable how usable this
is. Our certificate bundle is being phased out and what do we see in the
rest of the bundle? If there is any need in future we can always create
new module simple as this one.