diagnostics.sh: hide order numbers
Initial numbers are just for order of modules. They ensure that when we
run all modules and well as when we list them they are in specific
order. Printing them to user might be too confusing.

The only requirement is to ensure that even caller that selectively
chooses modules keeps the order but that can be ensured by keeping the
original print order without knowing the numbers.