Refactor and join few initial modules
Every module can be selectively chosen and is visible in (re)Foris. It
makes little to no sense to have router uname and other router
identification steps separate. In most cases we just want all of them
just to know router we are dealing with. In the end it creates clutter
in web interface.

This merges:
* serial number
* OS release
* uname
* uptime
* current date and time
into the one system-info module.

Date module after removal of current date and time was renamed to ntp.

Netstat module was renamed to sockets and uses newer 'ss' instead of
'netstat'. The 'ss' provides us with even more info.

Packages module was extended to debug opkg on top of just listing
packages. This is required not only to debug opkg bug also to compare
synchronization of configuration with updater.