Makefile: few improvements and fixes
This includes small hack to cover case when there is no change in po
file but due to msgmerge not updating modify time of that file the merge
is performed on every run.

This also makes sure that when update is run that it really updates all
messages not just those for change. This is intentional to make it
consistent with other repositories where update performs works
unconditionally.