web: add simple cgi scripts to generate diagnostics and export snapshots
The common way is that users can generate diagnostics as well as export
schnapshots in reForis but it is harder to get them once reForis is
crashing. This instead is little script that provides the same
functionality (although in not that nice coating) and works without
reForis. It covers situation when reForis does not work but Lighttpd
does and prevents need for users to go to shell to get diagnostics.